refactor: move market containers to market_sections dir
