+++ app/src/components/index.ts
@@ -1,7 +1,7 @@
 export { MarketWizardCreator } from './market/market_create/market_wizard_creator'
-export { MarketWizardCreatorContainer } from './market/market_create/market_wizard_creator_container'
-export { MarketViewContainer } from './market/market_profile/market_view_container'
+export { MarketWizardCreatorContainer } from '../pages/market_sections/market_wizard_creator_container'
+export { MarketViewContainer } from '../pages/market_sections/market_view_container'
 export { MarketPoolLiquidityContainer } from '../pages/market_sections/market_pool_liquidity_container'
 export { MarketSellContainer } from '../pages/market_sections/market_sell_container'
 export { MarketHome } from './market/market_list/market_home'
-export { MarketHomeContainer } from './market/market_list/market_home_container'
+export { MarketHomeContainer } from '../pages/market_sections/market_home_container'
